refactor: move mapper package to core module
