+++ xxl-job-admin-core/src/main/java/com/xxl/job/admin/core/mapper/XxlJobLogGlueMapper.java
@@ -1,6 +1,6 @@
-package com.xxl.job.admin.mapper;
+package com.xxl.job.admin.core.mapper;
 
-import com.xxl.job.admin.model.XxlJobLogGlue;
+import com.xxl.job.admin.core.model.XxlJobLogGlue;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -12,13 +12,13 @@ import java.util.List;
  */
 @Mapper
 public interface XxlJobLogGlueMapper {
-	
+
 	public int save(XxlJobLogGlue xxlJobLogGlue);
-	
+
 	public List<XxlJobLogGlue> findByJobId(@Param("jobId") int jobId);
 
 	public int removeOld(@Param("jobId") int jobId, @Param("limit") int limit);
 
 	public int deleteByJobId(@Param("jobId") int jobId);
-	
+
 }
