+++ xxl-job-admin-core/src/main/resources/mapper/XxlJobLockMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.xxl.job.admin.mapper.XxlJobLockMapper">
+<mapper namespace="com.xxl.job.admin.core.mapper.XxlJobLockMapper">
 
 	<select id="scheduleLock" resultType="java.lang.String" >
 		SELECT * FROM xxl_job_lock
