+++ xxl-job-admin-core/src/main/resources/mapper/XxlJobRegistryMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" 
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.xxl.job.admin.mapper.XxlJobRegistryMapper">
-	
-	<resultMap id="XxlJobRegistry" type="com.xxl.job.admin.model.XxlJobRegistry" >
+<mapper namespace="com.xxl.job.admin.core.mapper.XxlJobRegistryMapper">
+
+	<resultMap id="XxlJobRegistry" type="com.xxl.job.admin.core.model.XxlJobRegistry" >
 		<result column="id" property="id" />
 	    <result column="registry_group" property="registryGroup" />
 	    <result column="registry_key" property="registryKey" />
@@ -24,7 +24,7 @@
 		FROM xxl_job_registry AS t
 		WHERE t.update_time <![CDATA[ < ]]> DATE_ADD(#{nowTime},INTERVAL -#{timeout} SECOND)
 	</select>
-	
+
 	<delete id="removeDead" parameterType="java.lang.Integer" >
 		DELETE FROM xxl_job_registry
 		WHERE id in
