+++ xxl-job-admin-core/src/main/resources/mapper/XxlJobUserMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.xxl.job.admin.mapper.XxlJobUserMapper">
+<mapper namespace="com.xxl.job.admin.core.mapper.XxlJobUserMapper">
 
-	<resultMap id="XxlJobUser" type="com.xxl.job.admin.model.XxlJobUser" >
+	<resultMap id="XxlJobUser" type="com.xxl.job.admin.core.model.XxlJobUser" >
 		<result column="id" property="id" />
 		<result column="username" property="username" />
 	    <result column="password" property="password" />
@@ -61,7 +61,7 @@
 		WHERE t.id = #{id}
 	</select>
 
-	<insert id="save" parameterType="com.xxl.job.admin.model.XxlJobUser" useGeneratedKeys="true" keyProperty="id" >
+	<insert id="save" parameterType="com.xxl.job.admin.core.model.XxlJobUser" useGeneratedKeys="true" keyProperty="id" >
 		INSERT INTO xxl_job_user (
 			username,
 			password,
@@ -75,7 +75,7 @@
 		);
 	</insert>
 
-	<update id="update" parameterType="com.xxl.job.admin.model.XxlJobUser" >
+	<update id="update" parameterType="com.xxl.job.admin.core.model.XxlJobUser" >
 		UPDATE xxl_job_user
 		SET
 			<if test="password != null and password != ''">
